requirements.txt
fastapi
uvicorn
langchain
langchain-openai
sqlalchemy
pydantic
python-dotenv
requests
opensearch-py
sentence-transformers
